gendilsig: Implement readFileAlloc and use it when reading digest from file
Implement readFileAlloc that allocates a buffer for input data read from
file before reading the actual data into the buffer. Use this function in
gendilsig.c when reading data digest. Rename variables related to the digest
to have TBD (to-be-signed) in their names to prepare it for the next steps
where input data may also be the raw (unhashed) data that are to be signed
when using ML-DSA in pure mode.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>